feat(web,hasura): introduce standard jwt auth, GraphQL spec types, and WIP collab page content
Our Hasura instance is now set to public mode, which means it's impossible to hit directly without
being authenticated through supabase. We've also reset the config of the connectors to follow
graphql spec, and as such the generated ts types now adhere to that spec. Lastly, we've begun
working on an early WIP for the collab interface.